+++ payment_tipiregie/models/inherited_payment_transaction.py
@@ -23,11 +23,16 @@ class TipiRegieTransaction(models.Model):
     # region Fields declaration
     tipiregie_operation_identifier = fields.Char(
         string='Operation identifier',
-        help='Reference of the request of TX as stored in the acquirer database'
+        help='Reference of the request of TX as stored in the acquirer database',
     )
 
     tipiregie_return_url = fields.Char(
         string='Return URL',
+    )
+
+    tipiregie_sent_to_webservice = fields.Boolean(
+        string="Sent to tipiregie webservice",
+        default=False,
     )
 
     # endregion
@@ -39,6 +44,21 @@ class TipiRegieTransaction(models.Model):
     # endregion
 
     # region CRUD (overrides)
+    @api.model
+    def create(self, vals):
+        res = super(TipiRegieTransaction, self).create(vals)
+        if res.acquirer_id.provider == 'tipiregie':
+            prec = self.env['decimal.precision'].precision_get('Product Price')
+            email = res.partner_email
+            amount = int(float_round(res.amount * 100.0, prec))
+            reference = res.reference.replace('/', '  slash  ')
+            acquirer_reference = '%.15d' % int(uuid.uuid4().int % 899999999999999)
+            res.acquirer_reference = acquirer_reference
+            idop = res.acquirer_id.tipiregie_get_id_op_from_web_service(email, amount, reference, acquirer_reference)
+            res.tipiregie_operation_identifier = idop
+
+        return res
+
     # endregion
 
     # region Actions
@@ -126,23 +146,6 @@ class TipiRegieTransaction(models.Model):
                 'state_message': message,
             })
             return False
-
-    @api.multi
-    def tipiregie_generate_operation(self):
-        prec = self.env['decimal.precision'].precision_get('Product Price')
-        for tx in self:
-            email = tx.partner_id.email
-            price = int(float_round(tx.amount * 100.0, prec))
-            object = tx.reference.replace('/', '  slash  ')
-            acquirer_reference = tx.acquirer_reference or '%.15d' % int(uuid.uuid4().int % 899999999999999)
-            idop = tx.acquirer_id.tipiregie_get_id_op_from_web_service(email, price, object, acquirer_reference)
-
-            tx.write({
-                'acquirer_reference': acquirer_reference,
-                'tipiregie_operation_identifier': idop,
-            })
-
-        return
 
     @api.model
     def tipiregie_cron_check_draft_payment_transactions(self, options={}):