+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+ - [Fix] Check if payment transaction is already sent to TipiRégie webservice and recreate it in this case.
 
 ## [10.0.1.3.0] - 2019-07-19
 ### Added